materialFinder.js
This is node.js app that receives a string and returns all materials required for the value that was entered. 

It is currently unfinished, The returned list should be used to bounce back on the database table to retrieve all other information for the requirements i.e. the Item Name and other important details for front end exposure.