Basic economic data storage
This change creates a foundation for the data-objects that we will be using in order to transfer economic data from the service we are using towards the MySQL Database. 

The idea is to iterate through every item in the future economic_data table within MySQL and automate the call to the service where we then store the data values. This will run once daily in the future to reduce the demand that will be put on eve-central servers.

Note - Still need to remove non-tradeable goods from the table. 
Still have not designed a way to save historical data for trend analysis.

Data that we are worried about for the transfer are - 
1) Highest Buy Order Values
2) Lowest Sell Order Values
3) Profit Spread
4) Percentage Spread
5) Buy Volume (24 Hours)
6) Sell Volume (24 Hours)
7) All Volume (Buy + Sell (24 Hours))