Economic Extension
Updated the ecSpreadFinder to calculate and store new data from eve-central.
Prepared and cleaned out a new table which is a copy of invType called economicData. This will be the holding place of the economic data once queried and processed from eve-central. 

We are now calculating -
Buy Volume per Hour
Sell Volume per Hour
T.O Ratio (Turn-Over Ratio) which might be more indicative of demand instead of the turn-over rate.